Rename packages/server/server to packages/server/src
