+++ packages/server/src/schema/types/coffee_shop_type.js
